사용자정보등록, 사용자 정보조회,수정,삭제

+++ src/main/java/com/saehimit/shproject/entity/InboundManagement/InboundReceipt.java
@@ -24,7 +24,7 @@ public class InboundReceipt {
     private String productType; // 품목 타입
 
     @Column(nullable = false)
-    private int receivedQuantity; // 입고 수량
+    private Long receivedQuantity; // 입고 수량
 
     @Column(nullable = false)
     private LocalDateTime receivedDate; // 입고 날짜
